+++ Namo_SwiftUI/Projects/Feature/Category/Project.swift
@@ -47,7 +47,6 @@ let targets: [Target] = [
 		factory: .init(
 			infoPlist: .exampleAppDefault,
 			dependencies: [
-				.feature(interface: .Category),
 				.feature(implements: .Category)
 			]
 		)
