+++ Namo_SwiftUI/Projects/Feature/Friend/Interface/Sources/FeatureFriendInterfaceSample.swift
@@ -1,0 +1,7 @@
+//
+//  FeatureFriendInterfaceSample.swift
+//  FeatureFriendInterface
+//
+//  Created by 정현우 on 9/4/24.
+//
+
