--- Namo_SwiftUI/Projects/Feature/Friend/Interface/Sources/FeautreFriendInterfaceSample.swift
@@ -1,1 +1,0 @@
-
