+++ Namo_SwiftUI/Projects/Feature/Friend/Project.swift
@@ -47,7 +47,6 @@ let targets: [Target] = [
 		factory: .init(
 			infoPlist: .exampleAppDefault,
 			dependencies: [
-				.feature(interface: .Friend),
 				.feature(implements: .Friend)
 			]
 		)
