+++ Namo_SwiftUI/Projects/Feature/Friend/Sources/Components/FriendInfoPopup/FriendInfoPopupView.swift
@@ -1,0 +1,154 @@
+//
+//  FriendInfoPopupView.swift
+//  FeatureFriend
+//
+//  Created by 정현우 on 9/3/24.
+//
+
+import SwiftUI
+
+import ComposableArchitecture
+
+import SharedDesignSystem
+
+public struct FriendInfoPopupView: View {
+	let store: StoreOf<FriendInfoPopupStore>
+	
+	public init(store: StoreOf<FriendInfoPopupStore>) {
+		self.store = store
+	}
+
+	public var body: some View {
+		VStack(alignment: .leading, spacing: 0) {
+			friendInfoPopupImageAndNickname
+				.padding(.bottom, 20)
+			
+			friendInfoPopupDescriptionAndBirth
+				.padding(.bottom, 32)
+			
+			friendInfoPopupButtons
+		}
+		.padding(.horizontal, 30)
+		.padding(.top, 20)
+		.padding(.bottom, 26)
+	}
+	
+	private var friendInfoPopupImageAndNickname: some View {
+		HStack(spacing: 0) {
+			store.friend.image
+				.frame(width: 72, height: 72)
+				.clipShape(RoundedRectangle(cornerRadius: 15))
+				.padding(.trailing, 20)
+			
+			VStack(alignment: .leading, spacing: 6) {
+				Text(store.friend.nickname)
+					.font(.pretendard(.bold, size: 15))
+					.foregroundStyle(Color.mainText)
+					.lineLimit(1)
+				
+				Text("#\(store.friend.tag)")
+					.font(.pretendard(.bold, size: 15))
+					.foregroundStyle(Color.mainText)
+					.lineLimit(1)
+			}
+			
+			Spacer(minLength: 20)
+			
+			Button(
+				action: {
+//					store.send(.favoriteBtnTappedInInfo)
+				}, label: {
+					Image(asset: store.friend.isFavorite ? SharedDesignSystemAsset.Assets.icFavoriteFill : SharedDesignSystemAsset.Assets.icFavorite)
+						.resizable()
+						.frame(width: 28, height: 28)
+				}
+			)
+		}
+	}
+	
+	private var friendInfoPopupDescriptionAndBirth: some View {
+		VStack(alignment: .leading, spacing: 0) {
+			Text(store.friend.description)
+				.font(.pretendard(.regular, size: 15))
+				.foregroundStyle(Color.mainText)
+				.padding(.bottom, 16)
+			
+			HStack {
+				Text("이름")
+					.font(.pretendard(.bold, size: 15))
+					.foregroundStyle(Color.mainText)
+				
+				Spacer()
+				
+				Text(store.friend.name)
+					.font(.pretendard(.regular, size: 15))
+					.foregroundStyle(Color.mainText)
+			}
+			.padding(.bottom, 16)
+			
+			HStack {
+				Text("생일")
+					.font(.pretendard(.bold, size: 15))
+					.foregroundStyle(Color.mainText)
+				
+				Spacer()
+				
+				Text(store.friend.birthday)
+					.font(.pretendard(.regular, size: 15))
+					.foregroundStyle(Color.mainText)
+			}
+		}
+	}
+	
+	private var friendInfoPopupButtons: some View {
+		HStack {
+			Spacer(minLength: 0)
+			
+			Button(
+				action: {
+				
+				}, label: {
+					HStack(spacing: 12) {
+						Image(asset: SharedDesignSystemAsset.Assets.icTrashcan)
+							.resizable()
+							.frame(width: 20, height: 20)
+						
+						Text("일정")
+							.font(.pretendard(.regular, size: 15))
+							.foregroundColor(Color.colorBlack)
+					}
+					.frame(width: 120, height: 40)
+					.background(
+						Capsule()
+							.stroke(Color.colorBlack, lineWidth: 1)
+					)
+				}
+			)
+			.padding(.horizontal, 16)
+			
+			Button(
+				action: {
+					
+				}, label: {
+					HStack(spacing: 12) {
+						Image(asset: SharedDesignSystemAsset.Assets.icCalendar)
+							.resizable()
+							.frame(width: 20, height: 20)
+						
+						Text("삭제")
+							.font(.pretendard(.regular, size: 15))
+							.foregroundColor(Color.colorBlack)
+					}
+					.frame(width: 120, height: 40)
+					.background(
+						Capsule()
+							.stroke(Color.colorBlack, lineWidth: 1)
+					)
+				}
+			)
+			
+			Spacer(minLength: 0)
+			
+		}
+	}
+}
