--- Namo_SwiftUI/Projects/Feature/Friend/Sources/FeautreFriendSample.swift
@@ -1,1 +1,0 @@
-
